+++ scripts/stackstorm.js
@@ -354,12 +354,6 @@ module.exports = function(robot) {
       } else {
         data = req.body;
       }
-      // Special handler to try and figure out when a hipchat message
-      // is a whisper:
-      if (robot.adapterName === 'hipchat' && !data.whisper && data.channel.indexOf('@') > -1 ) {
-        data.whisper = true;
-        robot.logger.debug('Set whisper to true for hipchat message');
-      }
 
       postDataHandler.postData(data);
 
@@ -390,13 +384,6 @@ module.exports = function(robot) {
           data = JSON.parse(e.data).payload;
         } else {
           data = e.data;
-        }
-
-        // Special handler to try and figure out when a hipchat message
-        // is a whisper:
-        if (robot.adapterName === 'hipchat' && !data.whisper && data.channel.indexOf('@') > -1 ) {
-          data.whisper = true;
-          robot.logger.debug('Set whisper to true for hipchat message');
         }
 
         postDataHandler.postData(data);