+++ tests/test-formatdata.js
@@ -86,39 +86,6 @@ describe('SlackFormatter', function() {
 
 });
 
-describe('HipChatFormatter', function() {
-  var adapterName = 'hipchat';
-
-  it('should echo back for non-empty', function() {
-    var formatter = formatData.getFormatter(adapterName, null);
-    var o = formatter.formatData('DATA', null);
-    expect(o).to.be.an('string');
-    expect(o).to.equal('/code DATA');
-  });
-
-  it('should be an empty string for empty', function() {
-    var formatter = formatData.getFormatter(adapterName, null);
-    var o = formatter.formatData('', null);
-    expect(o).to.be.an('string');
-    expect(o).to.equal('');
-  });
-
-  it('should correctly format recepient', function() {
-    var formatter = formatData.getFormatter(adapterName, null);
-    env.HUBOT_HIPCHAT_JID = '234x_y234@conf.hipchat.com';
-    var o = formatter.formatRecepient('Estee');
-    expect(o).to.be.an('string');
-    expect(o).to.equal('234x_Estee@conf.hipchat.com');
-  });
-
-  it('should normalize command', function() {
-    var formatter = formatData.getFormatter(adapterName, null);
-    var o = formatter.normalizeCommand('run local "uname -a"');
-    expect(o).to.be.an('string');
-    expect(o).to.equal('run local "uname -a"');
-  });
-});
-
 describe('RocketChatFormatter', function() {
   var adapterName = 'rocketchat';